code cleaned up and began state machine

+++ SousVide/Inc/i2c-lcd.h
@@ -8,4 +8,4 @@ void lcd_send_data (char data);  // send data to the lcd
 
 void lcd_send_string (char *str);  // send string to the lcd
 
-void lcd_transmit(int length, char *data);
+void lcd_transmit(uint8_t length, char *data);
